+++ src/main/java/pl/asie/computronics/block/BlockPeripheral.java
@@ -22,20 +22,13 @@ import pl.asie.lib.block.BlockBase;
 })
 public abstract class BlockPeripheral extends BlockBase implements IComputronicsEnvironmentBlock, IBlockWithDocumentation {
 
-	protected boolean noNedoComputers = false;
-
 	public BlockPeripheral(String documentationName) {
 		super(Material.iron, Computronics.instance);
 		this.setCreativeTab(Computronics.tab);
 		this.documentationName = documentationName;
 	}
 
-	protected void setNoNedoComputers(boolean noNedoComputers) {
-		this.noNedoComputers = noNedoComputers;
-	}
-
 	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
 	public boolean onBlockActivated(World world, int x, int y, int z, EntityPlayer player, int a, float _x, float _y, float _z) {
 		TileEntity tile = world.getTileEntity(x, y, z);
 		if(tile instanceof TileEntityPeripheralBase && ((TileEntityPeripheralBase) tile).canBeColored()) {
@@ -44,15 +37,6 @@ public abstract class BlockPeripheral extends BlockBase implements IComputronics
 				((TileEntityPeripheralBase) tile).setColor(color.color);
 				world.markBlockForUpdate(x, y, z);
 				return true;
-			}
-		}
-		if(!world.isRemote && !noNedoComputers && player.isSneaking()) {
-			try {
-				Object ncinst = this.getClass().getClassLoader().loadClass("nedocomputers.NedoComputers").getField("instance").get(null);
-				player.openGui(ncinst, 2, world, x, y, z);
-				return true;
-			} catch(Exception e) {
-				e.printStackTrace();
 			}
 		}
 		return super.onBlockActivated(world, x, y, z, player, a, _x, _y, _z);