+++ src/main/java/pl/asie/computronics/block/BlockRadar.java
@@ -17,7 +17,6 @@ public class BlockRadar extends BlockMachineSidedIcon {
 		this.setCreativeTab(Computronics.tab);
 		this.setIconName("computronics:radar");
 		this.setBlockName("computronics.radar");
-		this.setNoNedoComputers(true);
 		this.setRotation(Rotation.NONE);
 	}