+++ src/main/java/pl/asie/computronics/integration/railcraft/tile/TileLocomotiveRelay.java
@@ -279,22 +279,4 @@ public class TileLocomotiveRelay extends TileEntityPeripheralBase {
 		}
 		return null;
 	}
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public boolean connectable(int side) {
-		return false;
-	}
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public short busRead(int addr) {
-		return 0;
-	}
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public void busWrite(int addr, short data) {
-
-	}
 }