+++ src/main/java/pl/asie/computronics/tile/TileCamera.java
@@ -99,33 +99,4 @@ public class TileCamera extends TileEntityPeripheralBase {
 		}
 		return null;
 	}
-
-	private float _nedo_xDir, _nedo_yDir;
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public short busRead(int addr) {
-		switch((addr & 0xFFFE)) {
-			case 4:
-				return ((short) (camera.getDistance() * 64));
-		}
-		return 0;
-	}
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public void busWrite(int addr, short data) {
-		float dataAsDir = (data / 32768.0F);
-		switch((addr & 0xFFFE)) {
-			case 0:
-				_nedo_xDir = dataAsDir;
-				break;
-			case 2:
-				_nedo_yDir = dataAsDir;
-				break;
-		}
-		if(addr < 4) {
-			camera.ray(worldObj, xCoord, yCoord, zCoord, getFacingDirection(), _nedo_xDir, _nedo_yDir);
-		}
-	}
 }