+++ src/main/java/pl/asie/computronics/tile/TileCipherBlock.java
@@ -179,62 +179,6 @@ public class TileCipherBlock extends TileEntityPeripheralBase implements IBundle
 		return null;
 	}
 
-	private short[] _nedo_cipher = new short[16];
-	private byte _nedo_status = 0;
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public short busRead(int addr) {
-		int a = ((addr & 0xFFFE)) >> 1;
-		if(a < 16) {
-			return _nedo_cipher[a];
-		} else if(a == 16) {
-			return _nedo_status;
-		} else if(a == 18) {
-			return (short) (isLocked ? 1 : 0);
-		} else {
-			return 0;
-		}
-	}
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public void busWrite(int addr, short data) {
-		if(cipher == null) {
-			return;
-		}
-
-		int a = ((addr & 0xFFFE)) >> 1;
-		if(a < 16) {
-			_nedo_cipher[a] = data;
-		} else if(a == 16) {
-			if(skey == null) {
-				updateKey();
-			}
-			try {
-				if(data == 1) { // Encrypt
-					cipher.init(Cipher.ENCRYPT_MODE, skey, new IvParameterSpec(iv));
-				} else if(data == 2) { // Decrypt
-					cipher.init(Cipher.DECRYPT_MODE, skey, new IvParameterSpec(iv));
-				}
-				byte[] _data = new byte[32];
-				for(int i = 0; i < 16; i++) {
-					_data[i * 2] = (byte) (_nedo_cipher[i] & 255);
-					_data[i * 2 + 1] = (byte) (_nedo_cipher[i] >> 8);
-				}
-				byte[] _out = cipher.doFinal(_data);
-				for(int i = 0; i < 16; i++) {
-					_nedo_cipher[i] = (short) (_out[i * 2] | (_out[i * 2 + 1] << 8));
-				}
-				_nedo_status = 0;
-			} catch(Exception e) {
-				_nedo_status = 1;
-			}
-		} else if(a == 18) {
-			isLocked = (data != 0);
-		}
-	}
-
 	private int bundledXORData;
 
 	private int getBundledXORKey() {