+++ src/main/java/pl/asie/computronics/tile/TileTapeDrive.java
@@ -622,53 +622,6 @@ public class TileTapeDrive extends TileEntityPeripheralBase implements IInventor
 		return null;
 	}
 
-	private int _nedo_lastSeek = 0;
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public short busRead(int addr) {
-		switch(addr & 0xFFFE) {
-			case 0:
-				return (short) state.getState().ordinal();
-			case 2:
-				return 0; // speed?
-			case 4:
-				return (short) state.soundVolume;
-			case 6:
-				return (state.getStorage() != null ? (short) (state.getStorage().getSize() / ItemTape.L_MINUTE) : 0);
-			case 8:
-				return (short) _nedo_lastSeek;
-			case 10:
-				return (state.getStorage() != null ? (short) state.getStorage().read(false) : 0);
-		}
-		return 0;
-	}
-
-	@Override
-	@Optional.Method(modid = Mods.NedoComputers)
-	public void busWrite(int addr, short data) {
-		switch(addr & 0xFFFE) {
-			case 0:
-				switchState(State.values()[data % State.values().length]);
-				break;
-			case 2:
-				break; // speed?
-			case 4:
-				state.soundVolume = Math.max(0, Math.min(data, 127));
-				break;
-			case 6:
-				break; // tape size is read-only!
-			case 8:
-				_nedo_lastSeek = state.getStorage().seek(data);
-				break;
-			case 10:
-				if(state.getStorage() != null) {
-					state.getStorage().write((byte) (data & 0xFF));
-				}
-				break;
-		}
-	}
-
 	@Override
 	public int getSourceId() {
 		return state.getId();