+++ changelog.md
@@ -1,5 +1,6 @@
 ==== 1.5.10 ====
 
+  * CHANGED: NedoComputers is no longer supported. Removed the EEPROM Reader and any kind of NedoComputers support there was.
   * FIXED: More crashes related to invalid crafting recipes.
 
 ==== 1.5.9 ====